+++ threedi_schema/migrations/versions/0229_clean_up.py
@@ -76,11 +76,32 @@ def remove_old_tables():
     remove_tables(remaining_v2_idx_tables + remaining_alembic)
 
 
+def clean_geometry_columns():
+    """ Remove columns referencing v2 in geometry_columns """
+    op.execute(sa.text("""
+            DELETE FROM geometry_columns WHERE f_table_name IN (
+                SELECT g.f_table_name FROM geometry_columns g
+                LEFT JOIN sqlite_master m ON g.f_table_name = m.name
+                WHERE m.name IS NULL AND g.f_table_name like "%v2%"
+            );
+        """))
+
+
+def clean_triggers():
+    """ Remove triggers referencing v2 tables """
+    connection = op.get_bind()
+    triggers = [item[0] for item in connection.execute(
+        sa.text("SELECT tbl_name FROM sqlite_master WHERE type='trigger' AND tbl_name LIKE '%v2%';")).fetchall()]
+    for trigger in triggers:
+        op.execute(f"DROP TRIGGER IF EXISTS {trigger};")
+
+
 def upgrade():
     remove_old_tables()
+    clean_geometry_columns()
+    clean_triggers()
     # Apply changing use_2d_rain and friction_averaging type to bool
     sync_orm_types_to_sqlite('model_settings')
-
 
 def downgrade():
     pass