+++ CHANGES.rst
@@ -6,8 +6,10 @@ Changelog of threedi-schema
 0.228.1 (unreleased)
 --------------------
 
+- Rename sqlite table "tags" to "tag"
 - Remove indices referring to removed tables in previous migrations
 - Make model_settings.use_2d_rain and model_settings.friction_averaging booleans
+- Remove columns referencing v2 in geometry_column